OrderControler, UserOrder, Ordermapper changes

+++ src/main/java/com/kodilla/ecommercee/service/OrderService.java
@@ -1,7 +1,9 @@
 package com.kodilla.ecommercee.service;
 
 import com.kodilla.ecommercee.entity.UserOrder;
+import com.kodilla.ecommercee.exception.OrderNotResolved;
 import com.kodilla.ecommercee.repository.OrderRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +13,7 @@ import java.util.Optional;
 @Service
 public class OrderService {
 
+    @Autowired
     private OrderRepository orderRepository;
 
     public List<UserOrder> getOrders() {
@@ -25,7 +28,13 @@ public class OrderService {
         return orderRepository.findById(orderId);
     }
 
-    public void deleteOrder(Long orderId) {
-        orderRepository.deleteById(orderId);
+    public void deleteOrder(Long orderId) throws OrderNotResolved {
+        if (!orderRepository.getOne(orderId).isResolved())
+        {
+            orderRepository.deleteById(orderId);
+        } else {
+            throw new OrderNotResolved();
+        }
+
     }
 }
